requirements.txt
Flask==3.0.3
azure-cosmos
azure-identity